+++ auth.json
@@ -1,1 +1,9 @@
-{}
+{
+    "services": {
+        "googlemapsportal": {
+            "service": "https://developers.google.com/maps/",
+            "username": "zhidou",
+            "key": "AIzaSyCUeg-W_secHpczmDVjngwSP7ATWaYxDOY"
+        }
+    }
+}
